+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- New `retries` field to `s3` input, to cap the number of download attempts made
+  on the same bucket item.
+
 ### 0.23.0 - 2018-08-06
 
 ### Added